+++ CHANGELOG.md
@@ -10,6 +10,16 @@
 
 ### 4. Hard deprecations
 
+#### Elixir
+
+  * [Calendar] `Calendar.ISO.day_of_week/3` is deprecated in favor of `Calendar.ISO.day_of_week/4`
+  * [Exception] `Exception.exception?/1` is deprecated in favor of `Kernel.is_exception/1`
+  * [Regex] `Regex.regex?/1` is deprecated in favor of `Kernel.is_struct/2`
+
+#### Logger
+
+  * [Logger] `Logger.warn/2` is deprecated in favor of `Logger.warning/2`
+
 ## v1.14
 
 The CHANGELOG for v1.14 releases can be found [in the v1.14 branch](https://github.com/elixir-lang/elixir/blob/v1.14/CHANGELOG.md).